+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Bugs fixed
 * Puppetresources does not fail tests for file sources starting with `file://`.
 * Escaped characters were not properly handled in the parser.
+* Properly catch division by 0 (!!!!).
 
 # v0.12.3 (2014/03/13)
 ## New featues